finished sdt, waiting fot the Pflag in tokenizerHelper and saving value at the end of the assembler file

+++ src/assemble_utils/instructions/instruction_defs.h
@@ -13,8 +13,8 @@ typedef struct {
     MNEMONIC mnemonic;
     cond_code cond;
     OPCODE opcode;
-    int Rn;
-    int Rd;
+    uint8_t Rn;
+    uint8_t Rd;
     uint32_t Operand2;
 } DataProcInstr_t;
 
@@ -23,10 +23,10 @@ typedef struct {
     cond_code cond;
     char A;
     char S;
-    int Rn;
-    int Rd;
-    int Rs;
-    int Rm;
+    uint8_t Rn;
+    uint8_t Rd;
+    uint8_t Rs;
+    uint8_t Rm;
 } MultiplyInstr_t;
 
 typedef struct {
@@ -35,8 +35,8 @@ typedef struct {
     char P;
     char U;
     char L;
-    int Rn;
-    int Rd;
+    uint8_t Rn;
+    uint8_t Rd;
     uint16_t offset;
 } SDTinstr_t;
